feature(movimentacão): Adiciona relatório de movimentação

+++ src/main/java/com/ti/estoque/services/ProductService.java
@@ -1,10 +1,15 @@
 package com.ti.estoque.services;
 
+import com.ti.estoque.enums.MovementType;
+import com.ti.estoque.exceptions.ResourceNotFoundException;
 import com.ti.estoque.models.Product;
+import com.ti.estoque.models.ProductMovement;
 import com.ti.estoque.repository.ProductRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -14,6 +19,7 @@ public class ProductService {
     private ProductRepository productRepository;
 
     public Product saveProduct(Product product) {
+        product.getMovements().add(new ProductMovement(product, product.getStockQuantity(), MovementType.ENTRY));
         return productRepository.save(product);
     }
 
@@ -21,9 +27,19 @@ public class ProductService {
         Product product = productRepository.findById(id).orElseThrow(() -> new RuntimeException("Produto não encontrado"));
         product.setName(productUpdated.getName());
         product.setCategory(productUpdated.getCategory());
-        product.setStockQuantity(productUpdated.getStockQuantity());
         product.setPrice(productUpdated.getPrice());
         product.setWarehouseLocation(productUpdated.getWarehouseLocation());
+        if (productUpdated.getStockQuantity() > product.getStockQuantity()) {
+            int quantity = productUpdated.getStockQuantity() - product.getStockQuantity();
+            product.getMovements().add(new ProductMovement(product, quantity, MovementType.ENTRY));
+        }
+        if (productUpdated.getStockQuantity() < product.getStockQuantity()) {
+            int quantity =  product.getStockQuantity() - productUpdated.getStockQuantity();
+
+            product.getMovements().add(new ProductMovement(product, quantity, MovementType.EXIT));
+        }
+        product.setStockQuantity(productUpdated.getStockQuantity());
+
         return productRepository.save(product);
     }
 
@@ -32,17 +48,40 @@ public class ProductService {
     }
 
     public Product findById(Long id) {
-        return productRepository.findById(id).orElseThrow(() -> new RuntimeException("Produto não encontrado"));
+        return productRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(String.format("Produto %s não encontrado", id)));
     }
 
     public void deleteById(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new EntityNotFoundException();
+        }
         productRepository.deleteById(id);
     }
 
-    public Product updateStock(Long id, Integer quantidade) {
-        Product produto = productRepository.findById(id).orElseThrow(() -> new RuntimeException("Produto não encontrado"));
-        produto.setStockQuantity(produto.getStockQuantity() + quantidade);
-        return productRepository.save(produto);
+    public Product updateStockEntry(Long id, Integer quantityChange) {
+        Product product = productRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Produto com ID " + id + " não encontrado."));
+
+        if (product.getStockQuantity() + quantityChange <= 0) {
+            throw new IllegalArgumentException("Quantidade tem que ser maior que 0.");
+        }
+
+        product.setStockQuantity(product.getStockQuantity() + quantityChange);
+        product.getMovements().add(new ProductMovement(product, quantityChange, MovementType.ENTRY));
+        return productRepository.save(product);
+    }
+
+    public Product updateStockExit(Long id, Integer quantityChange) {
+        Product product = productRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Produto com ID " + id + " não encontrado."));
+
+        if (product.getStockQuantity() + quantityChange <= 0) {
+            throw new IllegalArgumentException("Quantidade tem que ser maior que 0.");
+        }
+
+        product.setStockQuantity(product.getStockQuantity() + quantityChange);
+        product.getMovements().add(new ProductMovement(product, quantityChange, MovementType.EXIT));
+        return productRepository.save(product);
     }
 
     public List<Product> findByStockQuantityLessThan(int minimumQuantity) {
@@ -51,5 +90,9 @@ public class ProductService {
 
     public List<Product> findByStockQuantityGreaterThan(int maxQuantity) {
         return productRepository.findByStockQuantityGreaterThan(maxQuantity);
+    }
+
+    public List<Product> getProductsByDateRange(LocalDateTime startDate, LocalDateTime endDate) {
+        return productRepository.findBycreationDateBetween(startDate, endDate);
     }
 }
